feat(ui): add copyTextWithFallback utility for non-secure HTTP context
Copy buttons silently fail on LAN HTTP (e.g. http://192.168.x.x) because
navigator.clipboard is only available in secure contexts (HTTPS/localhost).

Created ui/src/lib/clipboard.ts with copyTextWithFallback() that:
- Uses navigator.clipboard.writeText() when available and isSecureContext
- Falls back to document.execCommand('copy') via textarea technique

Updated all direct navigator.clipboard calls:
- IssueChatThread.tsx (6 call sites → copyTextWithFallback)
- IssueContinuationHandoff.tsx (optional chaining → copyTextWithFallback)
- IssueDocumentsSection.tsx
- IssueWorkspaceCard.tsx
- IssueProperties.tsx
- CompanyInvites.tsx
- CompanySettings.tsx

CommentThread.tsx's local copyTextWithFallback removed in favor of the
shared utility. CopyText.tsx already had correct fallback - left as-is.

Fixes MIS-5367
