+++ ui/src/components/IssueContinuationHandoff.tsx
@@ -2,6 +2,7 @@ import { useCallback, useEffect, useRef, useState } from "react";
 import type { IssueDocument } from "@paperclipai/shared";
 import { ISSUE_CONTINUATION_SUMMARY_DOCUMENT_KEY } from "@paperclipai/shared";
 import { Button } from "@/components/ui/button";
+import { copyTextWithFallback } from "@/lib/clipboard";
 import { cn, relativeTime } from "../lib/utils";
 import { MarkdownBody } from "./MarkdownBody";
 import { Check, ChevronDown, ChevronRight, Copy, History } from "lucide-react";
@@ -40,7 +41,7 @@ export function IssueContinuationHandoff({
 
   const copyBody = useCallback(async () => {
     if (!document) return;
-    await navigator.clipboard?.writeText(document.body);
+    await copyTextWithFallback(document.body);
     setCopied(true);
     if (copiedTimerRef.current) {
       clearTimeout(copiedTimerRef.current);
